Completed task 1 TODO

+++ src/DataTypes.java
@@ -1,10 +1,11 @@
 import java.util.List;
 
 public class DataTypes {
-    // TODO TASK 1: fix this code so that it passes the test in DataTypesTest.java
+    // TASK 1: fix this code so that it passes the test in DataTypesTest.java
+    // changed int s to long s for the right variable type
     public static long sum(List<Integer> numbers) {
 
-        int s = 0;
+        long s = 0;
         // below is a "foreach" loop which iterates through numbers
         for (int x : numbers) {
             s += x;
